Added character movement system
The Character.cs script has been assigned to the Character prefab, which is used to control the character's movement. In the MoveTo() method we check by Raycast where on the plane the leader should go. This function checks which character is the leader and the rest are assigned the leader's Transform as a destination