Added characters stamina management system
The character's stamina decreases as it moves, its amount is generated randomly for each character at the beginning of the game. When stamina reaches zero, the character will stop for a predetermined time, after which the character will regenerate stamina to its initial value. When the character reaches the destination, his stamina will also regenerate