Added UI Information Panel
A panel has been added to the user interface that displays information about which character is currently the leader. And if we do not have a selected leader, it prompts the user to select one. The "IsLeaderAtDestination" field has been added to the "CharacterConfig" scriptable object, which allows to determine whether the leader has reached the destination point. It is needed to properly manage character animations. Also added several name changes in variables